Move ZeroGuid const from constants to guid file
