+++ packages/sdk/src/constants.ts
@@ -3,8 +3,6 @@
  * Licensed under the MIT License.
  */
 
-export const ZeroGuid = '00000000-0000-0000-0000-000000000000';
-
 /** @hidden */
 export const HTTPHeaders = {
 	SessionID: 'x-ms-mixed-reality-extension-sessionid',
