+++ packages/sdk/src/types/runtime/lookAt.ts
@@ -4,7 +4,7 @@
  */
 
 import { LookAtMode } from "../..";
-import { ZeroGuid } from "../../constants";
+import { ZeroGuidString as ZeroGuid } from "../..";
 
 export interface LookAtLike {
 	actorId: string;
